+++ src/document/Document.js
@@ -1,10 +1,11 @@
 const parse = require('./index');
 const sectionMap = require('./_sectionMap');
-const toMarkdown = require('../output/markdown');
-const toHtml = require('../output/html');
-const toJSON = require('../output/json');
-const toLatex = require('../output/latex');
+const toMarkdown = require('./toMarkdown');
+const toHtml = require('./toHtml');
+const toJSON = require('./toJSON');
+const toLatex = require('./toLatex');
 const setDefaults = require('../lib/setDefaults');
+const aliasList = require('../lib/aliases');
 
 const defaults = {
   infoboxes: true,
@@ -44,7 +45,10 @@ const methods = {
   reparse : function () {
     this.data = parse(this.wiki, this.options);
   },
-  wikitext : function() {
+  wikitext : function(str) {
+    if (str) {
+      this.wiki = str;
+    }
     return this.wiki;
   },
   isRedirect : function() {
@@ -108,6 +112,9 @@ const methods = {
   links : function(clue) {
     return sectionMap(this, 'links', clue);
   },
+  lists : function(clue) {
+    return sectionMap(this, 'lists', clue);
+  },
   tables : function(clue) {
     return sectionMap(this, 'tables', clue);
   },
@@ -123,9 +130,9 @@ const methods = {
   coordinates : function(clue) {
     return sectionMap(this, 'coordinates', clue);
   },
-  plaintext : function(options) {
+  text : function(options) {
     options = setDefaults(options, defaults);
-    let arr = this.sections().map(sec => sec.plaintext(options));
+    let arr = this.sections().map(sec => sec.text(options));
     return arr.join('\n\n');
   },
   markdown : function(options) {
@@ -152,6 +159,7 @@ const methods = {
       }
       console.log(indent + (sec.title() || '(Intro)'));
     });
+    return this;
   }
 };
 
@@ -172,13 +180,12 @@ plurals.forEach((fn) => {
 Object.keys(methods).forEach((k) => {
   Document.prototype[k] = methods[k];
 });
+//add alises, too
+Object.keys(aliasList).forEach((k) => {
+  Document.prototype[k] = methods[aliasList[k]];
+});
 //alias this one
-Document.prototype.toHTML = Document.prototype.html;
 Document.prototype.isDisambig = Document.prototype.isDisambiguation;
-Document.prototype.toJson = Document.prototype.json;
-Document.prototype.text = Document.prototype.plaintext;
 Document.prototype.references = Document.prototype.citations;
-Document.prototype.original = Document.prototype.wikitext;
-Document.prototype.wikiscript = Document.prototype.wikitext;
 
 module.exports = Document;