+++ src/section/toHtml.js
@@ -1,15 +1,4 @@
-const doSentence = require('./sentence');
-const doTable = require('./table');
-const makeImage = require('./image');
 
-const doList = (list) => {
-  let html = '<ul>\n';
-  list.forEach((o) => {
-    html += '  <li>' + o.text() + '</li>\n';
-  });
-  html += '<ul>\n';
-  return html;
-};
 
 const doSection = (section, options) => {
   let html = '';
@@ -23,22 +12,22 @@ const doSection = (section, options) => {
   if (options.images === true) {
     let imgs = section.images();
     if (imgs.length > 0) {
-      html += imgs.map((image) => makeImage(image)).join('\n');
+      html += imgs.map((image) => image.html(options)).join('\n');
       html += '\n';
     }
   }
   //make a html table
   if (options.tables === true) {
-    html += section.tables().map((t) => doTable(t, options)).join('\n');
+    html += section.tables().map((t) => t.html(options)).join('\n');
   }
   // //make a html bullet-list
   if (section.lists() && options.lists === true) {
-    html += section.lists().map((list) => doList(list, options)).join('\n');
+    html += section.lists().map((list) => list.html(options)).join('\n');
   }
   //finally, write the sentence text.
   if (options.sentences === true) {
     html += '  <div class="text">\n    ';
-    html += section.sentences().map((s) => doSentence(s, options)).join(' ');
+    html += section.sentences().map((s) => s.html(options)).join(' ');
     html += '\n  </div>\n';
   }
   return '<div class="section">\n' + html + '</div>\n';