+++ changelog.md
@@ -77,3 +77,8 @@
 * better disambiguation-page detection in english
 * remove wikitext from caption titles
 * support 3-level templates (whew!)
+## 5.0.0
+* new `Table` class and `List` classes
+* improved table-parser - generate name `col1` instead of `col-0`
+* support `options.verbose_template` for debugging
+* support recursive tables